+++ corehq/apps/app_manager/tests/__init__.py
@@ -1,6 +1,7 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
-from corehq.apps.app_manager.models import validate_property, CommentMixin
+from corehq.apps.app_manager.helpers.validators import validate_property
+from corehq.apps.app_manager.models import CommentMixin
 from corehq.apps.app_manager.util import is_valid_case_type, version_key
 from corehq.apps.app_manager.id_strings import _format_to_regex
 from corehq.apps.app_manager import xform_builder
